requirements.txt
streamlit==1.33.0
tensorflow-cpu==2.15.0
numpy==1.26.4
pillow==10.2.0
h5py==3.10.0
protobuf==4.25.3